+++ api/plugin-api/src/main/java/com/yanncebron/m68kplugin/browser/M68kBrowserPaneBase.java
@@ -58,7 +58,7 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 
 /**
- * @param <T> must implement {@link #equals(Object)} to keep current selection upon list model update
+ * @param <T> must implement {@link #equals(Object)} to keep the current selection upon list model update
  * @see M68kBrowserPaneFactory
  */
 @SuppressWarnings("UnstableApiUsage")
@@ -69,12 +69,6 @@ public abstract class M68kBrowserPaneBase<T> extends SimpleToolWindowPanel imple
    */
   @NonNls
   public static final String M68K_BROWSER_ITEM_LINK_PREFIX = "m68kBrowser://";
-
-  /**
-   * "Copy to clipboard" action.
-   */
-  @NonNls
-  public static final String M68K_BROWSER_COPY_DATA_LINK_PREFIX = "m68kBrowserCopy://";
 
   public static final Function<String, String> M68K_BROWSER_LINK_FUNCTION = link -> M68K_BROWSER_ITEM_LINK_PREFIX + StringUtil.substringBefore(link, ".md");